fixup! [front] refactor: share slash skill suggestion items
