+++ front/components/editor/extensions/input_bar/InputBarSlashSuggestionDropdown.tsx
@@ -1,13 +1,13 @@
+import {
+  getSkillSlashCommandItem,
+  matchesSlashCommandQuery,
+  sortSlashCommandMatches,
+} from "@app/components/editor/extensions/shared/SlashCommandSkillItems";
 import type {
   SlashCommand,
   SlashCommandDropdownRef,
 } from "@app/components/editor/extensions/skill_builder/SlashCommandDropdown";
 import { SlashCommandDropdown } from "@app/components/editor/extensions/skill_builder/SlashCommandDropdown";
-import {
-  getSkillSlashCommandItem,
-  matchesSlashCommandQuery,
-  sortSlashCommandMatches,
-} from "@app/components/editor/extensions/skill_builder/SlashCommandSkillItems";
 import {
   getMcpServerViewDescription,
   getMcpServerViewDisplayName,
