+++ front/components/editor/extensions/shared/SlashCommandSkillItems.ts
